+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 # All notable changes to this project will be documented in this file.
+## [Unreleased]
+### Added
+- MacOS support added
+- Single instance support added
+- MacOS mount plugin installer support implemented
+
+### Fixed
+- In the tray icon, the "Show App" option now correctly opens the app window. (Fixed)
+
 ## [beta-0.1.0] - 2024-12-05
 ### Added
 - Added a new feature to manage remotes with a user-friendly interface.